% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vignette_path.R
\name{build_vignette_path}
\alias{build_vignette_path}
\title{create raw user content path for a vignette}
\usage{
build_vignette_path(ghuser, ghrepo, branch, vname)
}
\arguments{
\item{ghuser}{user name}

\item{ghrepo}{repo name}

\item{branch}{branch name}

\item{vname}{vignette name}
}
\description{
create raw user content path for a vignette
}
