% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{desc}
\alias{desc}
\title{a list of 68 repository descriptions for packages in review in Nov 2025}
\format{
An object of class \code{list} of length 68.
}
\usage{
desc
}
\description{
a list of 68 repository descriptions for packages in review in Nov 2025
}
\note{
Each element has components `summary` and `URL` for a repository.

The summaries were produced by `summarize_submission` and
are based on analysis of DESCRIPTION and README.  This is an early
approach and `vig2data` might be preferred.  Or the two procedures
could be combines.
}
\examples{
data("desc", package="BiocRevSumm")
desc[[1]]
}
\keyword{datasets}
