% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_existing_branch.R
\name{find_existing_branch}
\alias{find_existing_branch}
\title{find branch name for an arbitrarily named submission to Bioconductor}
\usage{
find_existing_branch(
  user,
  repo,
  path,
  branches = c("main", "master", "devel", "origin")
)
}
\arguments{
\item{user}{GitHub user name}

\item{repo}{GitHub repo}

\item{path}{folder within repo}

\item{branches}{character vector of candidates}
}
\description{
find branch name for an arbitrarily named submission to Bioconductor
}
\examples{
find_existing_branch("vjcitn", "parody", "vignettes")
}
