% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_repo_urls.R
\name{get_issue_repo_urls}
\alias{get_issue_repo_urls}
\title{given an issue stream, find the first URL mentioned}
\usage{
get_issue_repo_urls(issues)
}
\arguments{
\item{issues}{a list of issues retrieved with `get_repo_issues`}
}
\description{
given an issue stream, find the first URL mentioned
}
