% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_repo_urls.R
\name{get_repo_issues}
\alias{get_repo_issues}
\title{given a github repo and token, enumerate all open issues}
\usage{
get_repo_issues(owner, repo, token, state = "open")
}
\arguments{
\item{owner}{organization name}

\item{repo}{repo name}

\item{token}{token such as GitHub PAT}

\item{state}{"open" by default, could be "all" but might hit rate limit}
}
\description{
given a github repo and token, enumerate all open issues
}
\examples{
if (nchar(tok <- Sys.getenv("GITHUB_TOKEN"))>0) {
 alli = get_repo_issues("bioconductor", "Contributions", tok)
 str(alli[[1]])
}
}
