% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_submission.R
\name{get_repo_metadata}
\alias{get_repo_metadata}
\title{produce a list with repository metadata, README, and DESCRIPTION for
a given repository}
\usage{
get_repo_metadata(owner, repo)
}
\arguments{
\item{owner}{organization}

\item{repo}{name of repository}
}
\description{
produce a list with repository metadata, README, and DESCRIPTION for
a given repository
}
\examples{
if (nchar(Sys.getenv("GITHUB_TOKEN"))>0) {
  lk = get_repo_metadata("bioconductor", "AnVILVRS")
}
}
