% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_github_dir.R
\name{list_github_dir}
\alias{list_github_dir}
\title{use GitHub API to list files in a repo folder}
\usage{
list_github_dir(user, repo, path, branch = "main")
}
\arguments{
\item{user}{GitHub user name string}

\item{repo}{string naming repo of interest}

\item{path}{folder to query}

\item{branch}{string naming branch}
}
\description{
use GitHub API to list files in a repo folder
}
\note{
Thank you perplexity.ai.
}
\examples{
list_github_dir("vjcitn", "parody", "vignettes")
}
