% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viglist_23_nov_2025}
\alias{viglist_23_nov_2025}
\title{URLs for vignettes of packages in review 22 Nov 2025}
\format{
character vector
}
\usage{
viglist_23_nov_2025
}
\description{
URLs for vignettes of packages in review 22 Nov 2025
}
\examples{
data("viglist_23_nov_2025", package="BiocRevSumm")
head(viglist_23_nov_2025)
}
\keyword{datasets}
